Create panna_kawaler_ii.vb
second version (GM6 table adjusted) of the macro that automates contracting parties marital status extraction. This version is adapted to all the new settings and rows and columns added to the GM6 table version. It's more automated than previous version and the "wspólność" property in the F column is now supported. I'm so proud of this second version.